requirements.txt
selenium
webdriver-manager
pytesseract
pillow
schedule
